refactor(ipc): 抽取 registerChromeIpc 到 ipc/register-ipc.js
main.js 的 34 个 IPC 通道注册(原 registerChromeIpc,约 555 行)整体迁出:

- 新增 ipc/register-ipc.js:registerIpc({ ipcMain, ctx, log }) 工厂,
  handler 主体逐行保持原状;依赖注入分两类:
  · 稳定引用(31 个动作函数 + 常量 + Set/Map + 模块)按引用传入
  · 可变状态(11 个模块级 let:mainWindow/serverProc/recovery/…)
    经 ctx.x() 访问器每次调用取当前值,与原先闭包读取等价;
    notifyOnTurnEnd/forceQuit 两处写入经 ctx.setX 提交
- main.js:registerChromeIpc 删除,替换为 registerAppIpc()(组装 ctx
  并调用 registerIpc),main.js 净减约 500 行;所有 ipcMain 注册清零
- 修正机械迁移中的三类问题:对象字面量简写属性坏点、局部 ctx 遮蔽
  外层 ctx(4 处改名 settingsCtx)、字符串字面量/注释内的 recovery 保留
- scripts/check-syntax.js 入口清单补 platform/ipc 新模块
- test 更新:recovery-integration 的 IPC 接线 pin 点指向 register-ipc.js;
  ipc-contracts 防漂移比对同时扫描 main.js + register-ipc.js

测试: 445 pass / 1 skip / 0 fail,git diff --check 通过
