+++ dsh-desktop/scripts/check-syntax.js
@@ -31,6 +31,9 @@ const entryFiles = [
   'preset-sync.js',
   'web-service-supervisor.js',
   'shutdown-coordinator.js',
+  'platform/process-tree.js',
+  'ipc/contracts.js',
+  'ipc/register-ipc.js',
 ];
 
 // 匹配「async/await 关键字与紧随其后的 function 声明之间被空行/注释行拆开」：
